feat(tvix): set up a simple command line parser for nix-store
We are going to have a 1:1 drop-in replacement for the old-style nix
tools, and this starts implementing the cli parser part.

The first step is to have a simple integration test suite that can
verify that we match the nix CLI.

clap is a super complicated parsing library, but looking through the
rest they are either too opinioated to be of use for us, or depend on
clap as implementation.

Change-Id: I4cf6051f3a4f782c3242fd0d2b9eab3fbe33d8ad
Reviewed-on: https://cl.tvl.fyi/c/depot/+/4756
Tested-by: BuildkiteCI
Reviewed-by: tazjin <tazjin@tvl.su>
Autosubmit: Profpatsch <mail@profpatsch.de>